refactor: split write policy path probes
